STM32 WATCHDOG : compilation issue with typed define

+++ targets/TARGET_STM/TARGET_STM32L1/device/stm32l1xx_hal_rcc_ex.h
@@ -58,7 +58,7 @@
  * @{
  */
 
-#define LSI_VALUE                  ((uint32_t)37000)  /* ~37kHz */
+#define LSI_VALUE                  (37000)  /* ~37kHz */
 
 #if defined(STM32L100xBA) || defined(STM32L151xBA) || defined(STM32L152xBA)\
  || defined(STM32L100xC) || defined(STM32L151xC) || defined(STM32L152xC)\
